requirements.txt
Flask
python-dotenv
Flask-Migrate
Flask-SQLAlchemy
uni-slugify
Flask-Cors
phonenumbers
Flask-Moment
Flask-WTF
Flask-Mail
Flask-Login
gunicorn
Flask-FlatPages
Bcrypt-Flask
email-validator
PyJWT
Flask-Minify
Flask-Caching
psycopg2-binary
Pillow
-i https://test.pypi.org/simple/ cinetpay-sdk==0.1.1
requests
celery
redis
flower
pyshorteners
httpx
